valgrind was also showing the heap was not completely  clear at the end of a run. These changes make sure the data on the heap (definitions for BN_two and BN_three) is cleared during the run.
Note
----
To run valgrind and check mem leaks, change the 'make' script as follows. For 
*each* line that starts:

    gcc -Wall -O3 --pedantic -Wno-variadic-macros \

copy/ paste this line and comment the first ot these out. Next change the 
uncommented line to

    gcc -Wall -ggdb --pedantic -Wno-variadic-macros \

and run ./make again

To run valgrind :

    valgrind -s --tool=memcheck --leak-check=yes --track-origins=yes ./es-demo-1

To switch back reverse the 'gcc' changes (abovei) or check out the original 
version of the 'make' script.